Extracted trimAndLowerCase utility to shared utility module
refs #9866
