+++ core/server/api/shared/serializers/input/all.js
@@ -1,17 +1,8 @@
 const debug = require('ghost-ignition').debug('api:shared:serializers:input:all');
 const _ = require('lodash');
+const utils = require('../../utils');
+
 const INTERNAL_OPTIONS = ['transacting', 'forUpdate'];
-
-const trimAndLowerCase = (params) => {
-    params = params || '';
-    if (_.isString(params)) {
-        params = params.split(',');
-    }
-
-    return params.map((item) => {
-        return item.trim().toLowerCase();
-    });
-};
 
 /**
  * Transform into model readable language.
@@ -21,17 +12,17 @@ module.exports = {
         debug('serialize all');
 
         if (frame.options.include) {
-            frame.options.withRelated = trimAndLowerCase(frame.options.include);
+            frame.options.withRelated = utils.options.trimAndLowerCase(frame.options.include);
             delete frame.options.include;
         }
 
         if (frame.options.fields) {
-            frame.options.columns = trimAndLowerCase(frame.options.fields);
+            frame.options.columns = utils.options.trimAndLowerCase(frame.options.fields);
             delete frame.options.fields;
         }
 
         if (frame.options.formats) {
-            frame.options.formats = trimAndLowerCase(frame.options.formats);
+            frame.options.formats = utils.options.trimAndLowerCase(frame.options.formats);
         }
 
         if (frame.options.formats && frame.options.columns) {
