+++ core/server/api/shared/utils/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    get options() {
+        return require('./options');
+    }
+};
